pyproject.toml
[project]
name = "jet_ml"
version = "0.1.0"
description = ""
authors = [
    {name = "Robert-Gomez-AI",email = "robertaidl@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas (>=2.2.3,<3.0.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "plotly (>=6.0.1,<7.0.0)",
    "matplotlib (>=3.10.1,<4.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "xgboost-cpu (>=3.0.0,<4.0.0)",
    "pydantic (>=2.11.3,<3.0.0)"
]

[tool.poetry]
packages = [{include = "jet_ml", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
